+++ changelog.md
@@ -3,3 +3,6 @@
 Migration information TODO.
 
 ## Breaking changes
+
+* The `StateTokenBusilder` and `StateTokenParser` classes were moved from `Stormpath.Owin.Abstractions` to `Stormpath.Owin.Middleware`.
+* The `LoginFormViewModelBuilder` class was moved from `Stormpath.Owin.Abstractions` to `Stormpath.Owin.Middleware`.